requirements.txt
kafka==1.3.5
requests-2.25.1
psycopg2==2.6.2